functions/tvs1/requirements.txt
appwrite
tradingview_screener
pandas
python-dotenv